backend/requirements.txt
plsfastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
aiohttp==3.9.1
requests==2.31.0
numpy==1.24.3
faiss-cpu==1.7.4
google-generativeai==0.3.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
web3==6.11.3
python-dateutil==2.8.2
eth-account==0.9.0
eth-utils==2.3.1

# LangChain and ML libraries
langchain==0.1.0
langchain-google-genai==0.0.5
langchain-community==0.0.10
chromadb==0.4.18
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2
scikit-learn==1.3.2
pandas==2.1.4
beautifulsoup4==4.12.2
lxml==4.9.3
feedparser==6.0.10
newsapi-python==0.2.7
pathway==0.8.0
tiktoken==0.5.2
openai==1.6.1
